Generalize IC/OBC extraction into model-agnostic engines
Split extract_forcings/obc.py and the old initial_condition.py into
model-agnostic GET/REGRID[/MERGE] engines (obc.py, ic.py) plus a
MOM6-specific callback layer (mom6.py) supplying the regrid step via
regional_mom6's Segment/experiment classes. driver.py's ic/bc dispatch
now goes through mom6.process_mom6_ic/process_mom6_obc.

Bumps the rm6 submodule pointer to pick up the newer experiment/segment
API (Grid/VGrid objects passed directly via hgrid_type/vgrid_type
instead of hgrid_path/vgrid_path strings; rotational_method removed
from regrid_velocity_tracers), which also required updating tides.py
and case.py's expt property to the new construction pattern.

Ported from work already done on the cice branch; this pulls out just
the generic engine split, without the CICE/WW3-specific pieces, so it
can land as its own layer beneath the forcing-configurator refactor.
